Modules/DevTools/requirements.txt
Cython~=0.29.23
python-gitlab~=2.8.0